+++ kong/plugins/http-log/schema.lua
@@ -1,5 +1,7 @@
 local typedefs = require "kong.db.schema.typedefs"
 local url = require "socket.url"
+local deprecation = require("kong.deprecation")
+
 
 return {
   name = "http-log",
@@ -8,15 +10,14 @@ return {
     { config = {
         type = "record",
         fields = {
-          -- NOTE: any field added here must be also included in the handler's get_queue_id method
           { http_endpoint = typedefs.url({ required = true, encrypted = true, referenceable = true }) }, -- encrypted = true is a Kong-Enterprise exclusive feature, does nothing in Kong CE
           { method = { type = "string", default = "POST", one_of = { "POST", "PUT", "PATCH" }, }, },
           { content_type = { type = "string", default = "application/json", one_of = { "application/json", "application/json; charset=utf-8" }, }, },
           { timeout = { type = "number", default = 10000 }, },
           { keepalive = { type = "number", default = 60000 }, },
-          { retry_count = { type = "integer", default = 10 }, },
-          { queue_size = { type = "integer", default = 1 }, },
-          { flush_timeout = { type = "number", default = 2 }, },
+          { retry_count = { type = "integer" }, },
+          { queue_size = { type = "integer" }, },
+          { flush_timeout = { type = "number" }, },
           { headers = {
             type = "map",
             keys = typedefs.header_name {
@@ -40,7 +41,29 @@ return {
               referenceable = true,
             },
           }},
+          { queue = typedefs.queue },
           { custom_fields_by_lua = typedefs.lua_code },
+        },
+
+        entity_checks = {
+          { custom_entity_check = {
+            field_sources = { "retry_count", "queue_size", "flush_timeout" },
+            fn = function(entity)
+              if entity.retry_count and entity.retry_count ~= 10 then
+                deprecation("http-log: config.retry_count no longer works, please use config.queue.max_retry_time instead",
+                            { after = "4.0", })
+              end
+              if entity.queue_size and entity.queue_size ~= 1 then
+                deprecation("http-log: config.queue_size no longer works, please use config.queue.max_batch_size instead",
+                            { after = "4.0", })
+              end
+              if entity.flush_timeout and entity.flush_timeout ~= 2 then
+                deprecation("http-log: config.flush_timeout no longer works, please use config.queue.max_coalescing_delay instead",
+                            { after = "4.0", })
+              end
+              return true
+            end
+          } },
         },
         custom_validator = function(config)
           -- check no double userinfo + authorization header